ENH: use fabio to perform fit2d mask saving and loading, with boolian switching to make certain masks end up in skbeam convention
